# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: CS2 Tracker API
  description: >
    HTTP API for the CS2 stats tracker: player profiles and careers, parsed
    match detail, weapon/map breakdowns, a leaderboard, demo ingest with job
    status, and Steam identity resolution.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /api/health:
    get:
      summary: Readiness probe (pings Postgres)
      responses:
        "200": { description: Healthy }
        "503": { description: Degraded (database unreachable) }
  /metrics:
    get:
      summary: Prometheus HTTP request counters
      responses:
        "200":
          description: Prometheus text exposition
          content:
            text/plain: {}
  /openapi.yaml:
    get:
      summary: This OpenAPI document
      responses:
        "200":
          description: OpenAPI spec
          content:
            application/yaml: {}

  /api/resolve:
    get:
      summary: Resolve a vanity name or SteamID64 to a SteamID64
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Resolved id
          content:
            application/json:
              schema:
                type: object
                properties: { steamId64: { type: string } }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { description: Vanity resolution needs a Steam API key }

  /api/leaderboard:
    get:
      summary: Top tracked players by rating
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 25 } }
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items: { $ref: "#/components/schemas/LeaderboardEntry" }

  /api/players/{steamid}:
    get:
      summary: Player profile (identity + rolling career)
      parameters: [{ $ref: "#/components/parameters/SteamID" }]
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerProfile" }
        "404": { $ref: "#/components/responses/NotFound" }

  /api/players/{steamid}/refresh:
    post:
      summary: Re-fetch identity from the Steam Web API (needs key)
      parameters: [{ $ref: "#/components/parameters/SteamID" }]
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerProfile" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { description: Steam API key not configured }

  /api/players/{steamid}/matches:
    get:
      summary: A player's recent matches with their per-match line
      parameters:
        - { $ref: "#/components/parameters/SteamID" }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items: { $ref: "#/components/schemas/PlayerMatchSummary" }

  /api/players/{steamid}/weapons:
    get:
      summary: Per-weapon kills and headshot %
      parameters:
        - { $ref: "#/components/parameters/SteamID" }
        - { name: limit, in: query, schema: { type: integer, default: 12 } }
      responses:
        "200":
          description: Weapon stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  weapons:
                    type: array
                    items: { $ref: "#/components/schemas/WeaponStat" }

  /api/players/{steamid}/maps:
    get:
      summary: Per-map career breakdown
      parameters: [{ $ref: "#/components/parameters/SteamID" }]
      responses:
        "200":
          description: Map stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  maps:
                    type: array
                    items: { $ref: "#/components/schemas/MapStat" }

  /api/players/{steamid}/leetify:
    get:
      summary: Live Leetify profile (fetched real-time; attributed to Leetify)
      parameters: [{ $ref: "#/components/parameters/SteamID" }]
      responses:
        "200": { description: Leetify profile (ratings, ranks, stats) }
        "404": { $ref: "#/components/responses/NotFound" }

  /api/players/{steamid}/faceit:
    get:
      summary: Live FACEIT profile (CS2 skill level, ELO, lifetime stats; needs FACEIT_API_KEY)
      parameters: [{ $ref: "#/components/parameters/SteamID" }]
      responses:
        "200": { description: FACEIT profile }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { description: FACEIT integration not configured }

  /api/players/{steamid}/steam-stats:
    get:
      summary: Raw Steam App 730 lifetime stats (needs key)
      parameters: [{ $ref: "#/components/parameters/SteamID" }]
      responses:
        "200": { description: Stat map }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { description: Steam API key not configured }

  /api/matches/{id}:
    get:
      summary: Full match detail (scoreboard + rounds)
      parameters: [{ $ref: "#/components/parameters/MatchID" }]
      responses:
        "200":
          description: Match detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchDetail" }
        "404": { $ref: "#/components/responses/NotFound" }

  /api/matches/{id}/kills:
    get:
      summary: Ordered killfeed for a match
      parameters: [{ $ref: "#/components/parameters/MatchID" }]
      responses:
        "200":
          description: Killfeed
          content:
            application/json:
              schema:
                type: object
                properties:
                  kills:
                    type: array
                    items: { $ref: "#/components/schemas/Kill" }

  /api/ingest/demo:
    post:
      summary: Enqueue a demo parse job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                demoPath: { type: string }
                demoUrl: { type: string }
                shareCode: { type: string }
                source: { type: string }
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string }
                  status: { type: string }
                  statusUrl: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }

  /api/jobs/{id}:
    get:
      summary: Parse-job status
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestJob" }
        "404": { $ref: "#/components/responses/NotFound" }

  /api/queue:
    get:
      summary: Pending parse-job count
      responses:
        "200":
          description: Depth
          content:
            application/json:
              schema:
                type: object
                properties: { depth: { type: integer, format: int64 } }

components:
  parameters:
    SteamID:
      name: steamid
      in: path
      required: true
      description: SteamID64 (17-digit)
      schema: { type: string }
    MatchID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
    Player:
      type: object
      properties:
        steamId64: { type: string }
        personaName: { type: string }
        avatarUrl: { type: string }
        profileUrl: { type: string }
        countryCode: { type: string }
    PlayerCareer:
      type: object
      properties:
        matches: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        roundsPlayed: { type: integer }
        kills: { type: integer, format: int64 }
        deaths: { type: integer, format: int64 }
        utilityDamage: { type: integer, format: int64 }
        enemiesFlashed: { type: integer, format: int64 }
        mvps: { type: integer, format: int64 }
        rating: { type: number }
        adr: { type: number }
        kastPct: { type: number }
        hsPct: { type: number }
        kd: { type: number }
        winRate: { type: number }
    PlayerProfile:
      type: object
      properties:
        player: { $ref: "#/components/schemas/Player" }
        career: { $ref: "#/components/schemas/PlayerCareer" }
    Match:
      type: object
      properties:
        id: { type: integer, format: int64 }
        map: { type: string }
        playedAt: { type: string, format: date-time }
        durationSeconds: { type: integer }
        roundsTotal: { type: integer }
        teamAScore: { type: integer }
        teamBScore: { type: integer }
        tickRate: { type: number }
    MatchPlayer:
      type: object
      properties:
        steamId64: { type: string }
        personaName: { type: string }
        startSide: { type: string, enum: ["CT", "T", ""] }
        kills: { type: integer }
        deaths: { type: integer }
        assists: { type: integer }
        adr: { type: number }
        kastPct: { type: number }
        hsPct: { type: number }
        rating: { type: number }
        won: { type: boolean }
    PlayerMatchSummary:
      type: object
      properties:
        match: { $ref: "#/components/schemas/Match" }
        line: { $ref: "#/components/schemas/MatchPlayer" }
    Round:
      type: object
      properties:
        number: { type: integer }
        winnerSide: { type: string, enum: ["CT", "T", ""] }
        endReason: { type: string }
        ctBuy: { type: string }
        tBuy: { type: string }
        ctEquipValue: { type: integer }
        tEquipValue: { type: integer }
    MatchDetail:
      type: object
      properties:
        match: { $ref: "#/components/schemas/Match" }
        players:
          type: array
          items: { $ref: "#/components/schemas/MatchPlayer" }
        rounds:
          type: array
          items: { $ref: "#/components/schemas/Round" }
    Kill:
      type: object
      properties:
        round: { type: integer }
        timeSeconds: { type: number }
        killerId: { type: string }
        victimId: { type: string }
        assisterId: { type: string }
        weapon: { type: string }
        headshot: { type: boolean }
        opening: { type: boolean }
        trade: { type: boolean }
    WeaponStat:
      type: object
      properties:
        weapon: { type: string }
        kills: { type: integer }
        headshots: { type: integer }
        hsPct: { type: number }
    MapStat:
      type: object
      properties:
        map: { type: string }
        matches: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        winRate: { type: number }
        rating: { type: number }
        adr: { type: number }
        kd: { type: number }
        hsPct: { type: number }
    LeaderboardEntry:
      type: object
      properties:
        steamId64: { type: string }
        personaName: { type: string }
        avatarUrl: { type: string }
        matches: { type: integer }
        rating: { type: number }
        kd: { type: number }
        adr: { type: number }
        winRate: { type: number }
    IngestJob:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        status: { type: string, enum: [queued, running, done, failed] }
        matchId: { type: integer, format: int64 }
        error: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }